Backend/requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
django-environ==0.11.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
reportlab==4.2.0
